requirements.txt
# GraphRAG项目依赖

# 核心依赖
openai>=1.0.0
pydantic>=2.0.0
networkx>=3.0
python-louvain>=0.16
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 文档处理
langchain>=0.1.0
tiktoken>=0.5.0
pymupdf>=1.23.0
python-docx>=0.8.11

# 图计算
igraph>=0.10.0
leidenalg>=0.10.0
community>=1.0.0

# 数据存储
sqlite3
sqlalchemy>=2.0.0
redis>=5.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 并发处理
asyncio
aiohttp>=3.8.0
concurrent.futures

# 评估和测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
rouge-score>=0.1.2

# 日志和监控
loguru>=0.7.0
tqdm>=4.65.0

# 可选依赖
# 用于可视化
matplotlib>=3.7.0
plotly>=5.15.0
pyvis>=0.3.2

# 用于更高级的NLP
spacy>=3.6.0
transformers>=4.30.0